Add conv2 forward as matrix multiplication
Created a copy of Conv2d class. Called it Conv2dWithLoops.
In Conv2d made a fully working forward implementation with
matrix multiplication instead of a loop.

As soon as he backward will be implemented as a matrix
multiplication, I won't use conversion of weights and biases
inside of calss. Those mthods will only be used to reinitialize
weigts and biases with weights and biases from other implementations.
May be it would be a nice idea to make the conversion to "regular"
shape and make a getter. So that users won't be scared :)