Fix Max Pool, complete np resnet
Fixed: MaxPool used to work properly only for padding = 0

Found out that in numpy_CNN training was so fast because Conv2d had
a fully matrix multiplication based backward prop (which gives correct
results only for some shapes of input and some hyperparameters).

Finished numpy version of resnet101. Forward and backward do not
cause exceptions. Don't know if it works properly because didn't train
it and didn't compare it with torch resnet. Need to add a method that
copies weights from torch implementation to compare them.