Free self.coverted_input in conv2d + cleanup
Deleted backward_as_mat_mul_unified. Not only it was an uglier version of backward_as_matrix_multiplication, but it fore some reason ran slower in all tests.

Now freeing memry in Conv2d backward (self.coverted_input). I trained a resnet101 on mnist and the model weighted over 5Gb. I think it is because we store self.coverted_input, which is a huge tensor.

Anyway it is still needed to code methods for saving to file and loading all parameters so that a user dosen't need to pickle a whole model.