Fix Conv2d backward
Conv2d backward uses matrix multiplication now. It works properly thus the semi-matmul approach was deleted. The propblem was that after in a conversion of input gradient after multiplication the intersections should be accumulated.

Was surprised to notice that backward works much slower than in previous commits (where the intersections were overwriting instead of accumulating). Apparently, the summation takes up a lot of time